+++ pkg/packages/packages_test.go
@@ -1,4 +1,4 @@
-package binaries
+package packages
 
 import (
 	"testing"
